async/spi: add helper methods to SpiDevice.
It's not possible to add them as default methods to `SpiDevice` itself,
but it's possible to do via an extension trait!

With this, you can avoid interacting with the horrible async closure hack in most
cases. You can now do `my_device.write(&buf).await`, just like with bare buses.