Make Cargo feature explicit
To keep consistent with all other crates, and for clarity, it is best to explicitly declare `defmt-03` as a feature of the `embedded-hal` crate. This way people can look at the `[features]` section rather than search through any optional dependencies.  In this case it's not that big of a deal, but it keeps the pattern.