Add Cursor type
This is a no_std variant of `std::io::Cursor`.

Fixes #631.

The implementation was generated by an LLM by copying the std impl and
adapting it. I (Zeeshan) reviewed the code and adjusted small bits. The
tests added are further guarantee that the code is doing what it's
supposed to do.